+++ trunk/ms-wind/MSWind/Scripts/LoginCharacter.cs
@@ -8,7 +8,7 @@ namespace MSWind.Scripts
         public LoginCharacter(Client client, int id)
         {
             PacketWriter Writer = new PacketWriter();
-            Writer.WriteShort(PacketOpcodes.sSelectCharacter);
+            Writer.WriteShort((short)PacketOpcodes.sSelectCharacter);
             Writer.WriteInt(client.Account.Characters[id].ID);
             client.SendPacket(Writer);
             // Program.mFrm.CharLoggedIn(id);
